requirements.txt
Django==5.1.4
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
mysqlclient==2.2.6
cryptography==44.0.0
python-dotenv==1.0.1
requests
django_redis